allow signed colors

+++ src/config.rs
@@ -124,7 +124,7 @@ impl StlConfig {
 /// DrawConfig is used to color histogram results
 #[derive(Clone, Serialize, Deserialize, Debug)]
 pub struct DrawConfig {
-    pub colors: Vec<[u8; 3]>,
+    pub colors: Vec<[i32; 3]>,
     pub powers: Vec<f64>,
     pub background_color: [f64; 3],
 }
